Add identity server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -52,12 +53,14 @@ public class AuctionsController : ControllerBase
         return Ok(_mapper.Map<AuctionDto>(auction));
     }
 
+    
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = _mapper.Map<Auction>(createAuctionDto);
-        //TODO: add current user as seller
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity?.Name ?? "Unknown user";
         
         _dbContext.Auctions.Add(auction);
         
@@ -72,6 +75,7 @@ public class AuctionsController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById), new {id = auction.Id}, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")] // api/auctions/{id}
     public async Task<ActionResult<AuctionDto>> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -80,6 +84,8 @@ public class AuctionsController : ControllerBase
             .FirstOrDefaultAsync(x => x.Id == id);
 
         if (auction == null) return NotFound();
+
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -96,11 +102,14 @@ public class AuctionsController : ControllerBase
         return Ok(_mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpDelete("{id}")] // api/auctions/{id}
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await _dbContext.Auctions.FindAsync(id);
         if (auction == null) return NotFound();
+
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         _dbContext.Auctions.Remove(auction);
